recover layer and field to use to choose colormaps
now it reverts to drawing old "image" fields directly and when given
a "value" field maps it through the lookup table and applies lighting.
Demonstrated the use of different color maps for different fields by
hardcoding the 'vBrownianVectorsX' field to use a hardcoded rainbow
color map. In principle we can have as many LUTs as we have value
fields now.