hook up lut to each field and fix a small bug
Specular lighting was off where normals pointed away from light. Now
it the specular component is 0,0,0 there (self shadowed).

The bigger change is that now each field has its own lookup table.
You can choose rainbow for salinity and spectral for temperature now
for example. The rendering widget edit section gives you a choice of
the available value fields (from the new FieldsModel) and the changes
you make in the rest of it pertain to the LUT associated with it.