Style fix in rendering widget, optimization to viz canvas widget.
Set min width of field name dropdown so that it lines up with the
left edge of the current value text box and color box.  Also set
the lookup table for a field code in the callback of the vis canvas
widget light so that we do not have to create them with every call
to draw a frame.