Support layer_color_by in info.json, but do not require it.
This adds the ability to restrict the color by options which show
up in the pipeline widget by adding a layer_color_by section to the
info.json.  Also, any field that shows up in a section within this
new optional layer_color_by will be added to the fields list in the
rendering widget.

If the info.json does not have this section, then all the old
behavior is preserved.